fzgpumodules: widen numeric options to fix Python boxing bug, port remaining FZGM stages, add rebuilt diagnostic metric and regression tests
- pressio_options::get() requires an exact type match and libpressio's
  Python layer always boxes int->int64_t, float->double; every plugin
  numeric option was narrower than that and so was silently never set
  from Python. Widen all affected fields (quant_radius, outlier_capacity,
  chunk_size, nbits, element_width, memory_multiplier, num_streams, etc.)
  to int64_t/double across stage_bitpack.h, stage_bitshuffle.h,
  stage_lorenzo.h, stage_quantizer.h, stage_rze.h, and fzgpumodules.cc.
- Add "prel" (pseudo-relative) error bound mode.
- Add fusion policy option (off/auto/aggressive) as a first-class,
  documented option.
- Add centering, dither, dither_seed, dither_strength, linear_mode to
  the quantizer; linear_mode now surfaces the engine's overflow guard
  (round() exceeding the signed code range raises instead of wrapping).
- Port the remaining ~18 unwired FZGM stages to typed Python handlers:
  tiled_lorenzo, adaptive_bitpack, adaptive_lorenzo, rre, rare, raze,
  clog, hclog, gpulz, tupl, ans, adm, bitplane_rze, szx, szp, huffman,
  log_transform, ginterp, roibin_split.
- Add fzgpumodules:rebuilt metric so Python callers can tell whether an
  encode() call paid pipeline-(re)build cost, for timing-loop hygiene.
- test_fzgm.cc: ~27 new regression tests covering fusion, error-bound
  modes, per-stage options, and the new stage types; fix 13 pre-existing
  and new test literals whose type didn't exactly match the (now wider)
  C++ field, which silently defeated the option set the same way the
  Python boxing bug did.
- test_fzgm_python.py: new Python-level regression suite exercising the
  actual int/float boxing path that test_fzgm.cc's C++ literals cannot.

+++ src/plugins/compressors/fzgpumodules/stage_bitpack.h
@@ -4,7 +4,10 @@
 namespace libpressio { namespace fzgpumodules { namespace fzgpumodules_ns {
 
 struct BitpackParams {
-    int nbits = 0;  // 0 = stage default (full width); set to power-of-2 to save bits
+    // int64_t (not int): pressio_options::get() requires an exact type match
+    // and libpressio's Python layer always boxes int->int64_t, so a narrower
+    // field here would silently never be set.
+    int64_t nbits = 0;  // 0 = stage default (full width); set to power-of-2 to save bits
     bool operator==(const BitpackParams& o) const { return nbits == o.nbits; }
     bool operator!=(const BitpackParams& o) const { return !(*this == o); }
 };
@@ -22,17 +25,17 @@ public:
 
         if(t == "uint8") {
             auto* s = ctx.pipeline.addStage<fz::BitpackStage<uint8_t>>();
-            if(p.nbits > 0) s->setNBits(p.nbits);
+            if(p.nbits > 0) s->setNBits(static_cast<uint8_t>(p.nbits));
             return s;
         }
         if(t == "uint16") {
             auto* s = ctx.pipeline.addStage<fz::BitpackStage<uint16_t>>();
-            if(p.nbits > 0) s->setNBits(p.nbits);
+            if(p.nbits > 0) s->setNBits(static_cast<uint8_t>(p.nbits));
             return s;
         }
         if(t == "uint32") {
             auto* s = ctx.pipeline.addStage<fz::BitpackStage<uint32_t>>();
-            if(p.nbits > 0) s->setNBits(p.nbits);
+            if(p.nbits > 0) s->setNBits(static_cast<uint8_t>(p.nbits));
             return s;
         }
         throw std::runtime_error("Unsupported bitpack type: " + t);
